markused: fix generic fn instantiation dedup key collision for duplicate module names
When multiple modules share the same name (e.g., both declare 'module common'
in different directories) and define structs with identical short names, the
walker's fn_generic_types_key used type_to_str() for deduplication. Type names
from different same-named modules could produce identical keys, causing
subsequent generic instantiations to be silently dropped.

Fix by using u32(typ).str() (the raw Type ID) instead of type_to_str,
guaranteeing unique keys per concrete type regardless of module path.

Adds project_issue_27398 with 100 module pairs to reliably reproduce the bug.

Fixes #27398.

+++ vlib/v/markused/walker.v
@@ -510,7 +510,7 @@ fn (mut w Walker) record_used_fn_generic_types(fkey string, concrete_types []ast
 fn (w &Walker) fn_generic_types_key(fkey string, concrete_types []ast.Type) string {
 	mut parts := []string{cap: concrete_types.len}
 	for typ in concrete_types {
-		parts << typ.str()
+		parts << u32(typ).str()
 	}
 	return '${fkey}:${parts.join('|')}'
 }
